requirements.txt
async-timeout==4.0.2
cachelib==0.10.2
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
colorama==0.4.6
Flask==2.2.2
Flask-Session==0.4.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
python-dotenv==1.0.0
redis==4.5.1
requests==2.28.2
six==1.16.0
spotipy==2.20.0
urllib3==1.26.14
Werkzeug==2.2.3
